+++ src/main/java/com/easypost/exception/API/EncodingError.java
@@ -1,8 +1,8 @@
 package com.easypost.exception.API;
 
-import com.easypost.exception.EasyPostException;
+import com.easypost.exception.APIException;
 
-public class EncodingError extends EasyPostException {
+public class EncodingError extends APIException {
     /**
      * EncodingError constructor.
      *