+++ src/main/java/com/easypost/exception/API/ForbiddenError.java
@@ -3,9 +3,9 @@ package com.easypost.exception.API;
 import java.util.List;
 
 import com.easypost.model.Error;
-import com.easypost.exception.EasyPostException;
+import com.easypost.exception.APIException;
 
-public class ForbiddenError extends EasyPostException {
+public class ForbiddenError extends APIException {
     /**
      * ForbiddenError constructor.
      *