+++ src/main/java/com/easypost/exception/General/SignatureVerificationError.java
@@ -11,14 +11,4 @@ public class SignatureVerificationError extends EasyPostException {
     public SignatureVerificationError(final String message) {
         super(message);
     }
-
-    /**
-     * SignatureVerificationError constructor with nested exception.
-     *
-     * @param message the exception message
-     * @param e the nested exception
-     */
-    public SignatureVerificationError(final String message, Throwable e) {
-        super(message, e);
-    }
 }