+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [FIXED] Typo, `ServiceUnavailablError` is now `ServiceUnavailableError`
 - [FIXED] Typo: `Smartrate` is now `SmartRate`, `apikeys` is now `apiKey`, `scanform` is now `scanForm`
 - [FIXED] Move `validateWebook`, `findLowestSmartrate`, and `getLowestSmartRate` from Services to Utilities since it does not required an EasyPost client object
+- [ADDED] Add a root level `APIException` class for all HTTP exceptions
 
 ## v6.0.0-rc1 (2022-11-30)